requirements.txt
anthropic>=0.92.0
openai>=1.54.0
yt-dlp>=2024.11.0
requests>=2.32.0
resend>=2.5.0
jinja2>=3.1.0
gspread>=6.1.0
google-auth>=2.30.0
